requirements.txt
python-dateutil==2.8.2
kafka-python==2.0.2
faust-streaming~=0.10.0
requests~=2.28.1
Flask~=2.2.2
pandas~=1.5.1
scipy~=1.10.1
setuptools~=60.2.0
streamlit~=1.16.0